feat(db): support standalone MoonBoard sync daemon
