+++ packages/db/drizzle/meta/_journal.json
@@ -1548,6 +1548,13 @@
       "when": 1788736783044,
       "tag": "0220_climb_reports_hide",
       "breakpoints": true
+    },
+    {
+      "idx": 221,
+      "version": "7",
+      "when": 1788754031670,
+      "tag": "0221_moonboard_sync_daemon",
+      "breakpoints": true
     }
   ]
 }
